fix(emulator): split ADB bridge tests to stay under max-lines
emulator-bridge.test.ts was 958 counted lines (limit 800). Move the ADB
network-device and RuntimeEmulatorCommands ADB cases into
emulator-bridge-adb.test.ts so static analysis can pass.
