+++ src/main/emulator/emulator-bridge.test.ts
@@ -21,8 +21,7 @@ const {
   listServeSimHelperProcessesForDeviceMock: vi.fn(async (): Promise<ServeSimHelperProcess[]> => []),
   shutdownSimulatorDeviceMock: vi.fn(async () => {}),
   netFetchMock: vi.fn(),
-  // Default null: the iOS-focused tests below keep the android backend inert.
-  // A dedicated test overrides this to a real SDK to prove the emu-kill guard.
+  // Default null: keep the android backend inert in these iOS-focused tests.
   discoverAndroidSdkFromHostMock: vi.fn((): unknown => null),
   androidCommandRunnerMock: vi.fn(async (_binary: string, _args: readonly string[]) => ({
     stdout: '',
@@ -61,17 +60,12 @@ vi.mock('./simulator-app-visibility', () => ({
   hideNativeSimulatorApp: hideNativeSimulatorAppMock
 }))
 
-// Keep the Android backend inert in these iOS-focused tests (no host SDK, no adb I/O)
-// by default; one test overrides discoverAndroidSdkFromHostMock to exercise the
-// android backend's emu-kill guard end-to-end through the bridge.
+// Keep the Android backend inert in these iOS-focused tests (no host SDK, no adb I/O).
 vi.mock('./android/android-sdk-host-discovery', () => ({
   discoverAndroidSdkFromHost: discoverAndroidSdkFromHostMock,
   setConfiguredAndroidSdkPath: () => {}
 }))
 
-// EmulatorBridge always constructs a real AndroidEmulatorBackend with the default
-// (execFile-based) command runner; swap it for a recording mock so the emu-kill
-// guard test can assert on the exact adb invocations without spawning a process.
 vi.mock('./android/android-command-runner', async (importOriginal) => {
   const actual = (await importOriginal()) as Record<string, unknown>
   return { ...actual, execFileAndroidCommandRunner: androidCommandRunnerMock }
@@ -217,166 +211,11 @@ describe('EmulatorBridge helper ownership', () => {
     expect(bridge.getActiveForWorktree('wt-external')).toBeNull()
   })
 
-  it('never sends emu kill to a TCP or USB android device during destroyAllSessions/app-quit', async () => {
-    discoverAndroidSdkFromHostMock.mockReturnValue({
-      sdkRoot: '/sdk',
-      adb: '/sdk/adb',
-      avdTools: null
-    })
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary === '/sdk/adb' && args.join(' ') === 'devices -l') {
-        return {
-          stdout:
-            'List of devices attached\n' +
-            '127.0.0.1:5555\tdevice\n' +
-            'cloud.internal:5555\tdevice\n' +
-            'R58N123ABC\tdevice',
-          stderr: '',
-          code: 0
-        }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-    bridge.registerActiveEmulator('wt-tcp', session('127.0.0.1:5555'), {
-      managed: true,
-      backend: 'android'
-    })
-    bridge.registerActiveEmulator('wt-cloud', session('cloud.internal:5555'), {
-      managed: true,
-      backend: 'android'
-    })
-    bridge.registerActiveEmulator('wt-usb', session('R58N123ABC'), {
-      managed: true,
-      backend: 'android'
-    })
-
-    await bridge.destroyAllSessions()
-
-    for (const call of androidCommandRunnerMock.mock.calls) {
-      expect(call[1].join(' ')).not.toContain('emu kill')
-    }
-    expect(bridge.getActiveForWorktree('wt-tcp')).toBeNull()
-    expect(bridge.getActiveForWorktree('wt-cloud')).toBeNull()
-    expect(bridge.getActiveForWorktree('wt-usb')).toBeNull()
-
-    // Re-register and drive the same guard through the app-quit path, which is
-    // the hard rule's other named entry point (onAppQuit just calls destroyAllSessions).
-    androidCommandRunnerMock.mockClear()
-    bridge.registerActiveEmulator('wt-tcp', session('127.0.0.1:5555'), {
-      managed: true,
-      backend: 'android'
-    })
-
-    await bridge.onAppQuit()
-
-    for (const call of androidCommandRunnerMock.mock.calls) {
-      expect(call[1].join(' ')).not.toContain('emu kill')
-    }
-  })
-
-  it('stops the scrcpy helper and clears the session registry before adb disconnect', async () => {
-    const ADDRESS = '127.0.0.1:5555'
-    discoverAndroidSdkFromHostMock.mockReturnValue({
-      sdkRoot: '/sdk',
-      adb: '/sdk/adb',
-      avdTools: null
-    })
-    let disconnected = false
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary === '/sdk/adb' && args.join(' ') === 'devices -l') {
-        return {
-          stdout: disconnected
-            ? 'List of devices attached'
-            : `List of devices attached\n${ADDRESS}\tdevice`,
-          stderr: '',
-          code: 0
-        }
-      }
-      if (binary === '/sdk/adb' && args[0] === 'disconnect') {
-        disconnected = true
-        return { stdout: `disconnected ${ADDRESS}`, stderr: '', code: 0 }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-    bridge.registerActiveEmulator('wt-1', session(ADDRESS), { managed: true, backend: 'android' })
-
-    const status = await bridge.adbDisconnect(ADDRESS)
-
-    expect(status).toEqual({ state: 'disconnected', address: ADDRESS, serial: null })
-    expect(bridge.getActiveForWorktree('wt-1')).toBeNull()
-    // The orphan-forward cleanup (part of stopHelperForDevice) must be issued
-    // strictly before `adb disconnect` — tearing down the stream before the
-    // connection it rides on.
-    const argLists = androidCommandRunnerMock.mock.calls.map((call) => call[1] as string[])
-    const forwardIndex = argLists.findIndex((args) => args.includes('forward'))
-    const disconnectIndex = argLists.findIndex((args) => args[0] === 'disconnect')
-    expect(forwardIndex).toBeGreaterThanOrEqual(0)
-    expect(disconnectIndex).toBeGreaterThan(forwardIndex)
-  })
-
-  it('never issues adb disconnect on pane-close (stopActiveForWorktree) or app quit', async () => {
-    const ADDRESS = '127.0.0.1:5555'
-    discoverAndroidSdkFromHostMock.mockReturnValue({
-      sdkRoot: '/sdk',
-      adb: '/sdk/adb',
-      avdTools: null
-    })
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary === '/sdk/adb' && args.join(' ') === 'devices -l') {
-        return {
-          stdout: `List of devices attached\n${ADDRESS}\tdevice`,
-          stderr: '',
-          code: 0
-        }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-    bridge.registerActiveEmulator('wt-1', session(ADDRESS), { managed: true, backend: 'android' })
-
-    await bridge.stopActiveForWorktree('wt-1', { shutdownDevice: true })
-    for (const call of androidCommandRunnerMock.mock.calls) {
-      expect(call[1][0]).not.toBe('disconnect')
-    }
-
-    androidCommandRunnerMock.mockClear()
-    bridge.registerActiveEmulator('wt-2', session(ADDRESS), { managed: true, backend: 'android' })
-    await bridge.onAppQuit()
-    for (const call of androidCommandRunnerMock.mock.calls) {
-      expect(call[1][0]).not.toBe('disconnect')
-    }
-  })
-
   it('rejects a capability the resolved backend does not support', async () => {
     const bridge = new EmulatorBridge()
     // device-1 resolves to the iOS backend, which advertises no explicit-verb caps.
     await expect(
       bridge.runCapability('install', { device: 'device-1' }, async () => 'unused')
-    ).rejects.toMatchObject({ code: 'emulator_unsupported' })
-  })
-
-  it('routes an offline ADB network address to the android backend, not the darwin/iOS fallback', async () => {
-    const bridge = new EmulatorBridge()
-    // iOS has no `install` capability, so if this routed to iOS (the pre-fix
-    // fallback on darwin) it would reject with emulator_unsupported instead.
-    const result = await bridge.runCapability(
-      'install',
-      { device: '127.0.0.1:5555' },
-      async () => 'routed-to-android'
-    )
-    expect(result).toBe('routed-to-android')
-    expect(listSimulatorDevicesMock).not.toHaveBeenCalled()
-    expect(execServeSimCommandMock).not.toHaveBeenCalled()
-  })
-
-  it('keeps the existing host-platform fallback for an unrecognized non-network identifier', async () => {
-    const bridge = new EmulatorBridge()
-    // 'unknown-device' is not host:port shaped, so it still falls through to the
-    // darwin-primary backend (iOS), matching pre-existing fallback behavior.
-    await expect(
-      bridge.runCapability('install', { device: 'unknown-device' }, async () => 'unused')
     ).rejects.toMatchObject({ code: 'emulator_unsupported' })
   })
 
@@ -944,164 +783,5 @@ describe('RuntimeEmulatorCommands attach lifecycle', () => {
       ['--detach', '-q', 'device-iphone'],
       { json: true }
     )
-  })
-})
-
-describe('RuntimeEmulatorCommands ADB device connection', () => {
-  const ADDRESS = '192.168.1.50:5555'
-
-  beforeEach(() => {
-    discoverAndroidSdkFromHostMock.mockReset()
-    discoverAndroidSdkFromHostMock.mockReturnValue({
-      sdkRoot: '/sdk',
-      adb: '/sdk/adb',
-      avdTools: null
-    })
-    androidCommandRunnerMock.mockReset()
-    androidCommandRunnerMock.mockImplementation(async () => ({ stdout: '', stderr: '', code: 0 }))
-  })
-
-  function commands(bridge: EmulatorBridge): RuntimeEmulatorCommands {
-    return new RuntimeEmulatorCommands({
-      getEmulatorBridge: () => bridge,
-      resolveEmulatorWorkspaceId: vi.fn(async () => 'wt-1'),
-      resolveEmulatorCleanupWorkspaceId: vi.fn(async () => 'wt-1'),
-      getAuthoritativeWindow: () => ({ webContents: { send: vi.fn() } }) as never,
-      getSettings: () => ({
-        mobileEmulatorEnabled: true,
-        mobileEmulatorDefaultDeviceUdid: null
-      })
-    })
-  }
-
-  it('delegates a successful connect to the bridge and returns its status verbatim', async () => {
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary === '/sdk/adb' && args[0] === 'connect') {
-        return { stdout: `connected to ${ADDRESS}`, stderr: '', code: 0 }
-      }
-      if (binary === '/sdk/adb' && args.join(' ') === 'devices -l') {
-        return { stdout: `List of devices attached\n${ADDRESS}\tdevice`, stderr: '', code: 0 }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-
-    const status = await commands(bridge).emulatorAdbConnect({ address: ADDRESS, worktree: 'wt-1' })
-
-    expect(status).toEqual({ state: 'connected', address: ADDRESS, serial: ADDRESS })
-  })
-
-  it('rejects a grammar-invalid address with emulator_adb_address_invalid, without touching adb', async () => {
-    const bridge = new EmulatorBridge()
-
-    await expect(
-      commands(bridge).emulatorAdbConnect({ address: 'not-an-address' })
-    ).rejects.toMatchObject({ code: 'emulator_adb_address_invalid' })
-    expect(androidCommandRunnerMock).not.toHaveBeenCalled()
-  })
-
-  it('surfaces an unauthorized status (not a thrown error) when the device needs authorization', async () => {
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary === '/sdk/adb' && args[0] === 'connect') {
-        return { stdout: `connected to ${ADDRESS}`, stderr: '', code: 0 }
-      }
-      if (binary === '/sdk/adb' && args.join(' ') === 'devices -l') {
-        return { stdout: `List of devices attached\n${ADDRESS}\tunauthorized`, stderr: '', code: 0 }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-
-    const status = await commands(bridge).emulatorAdbConnect({ address: ADDRESS })
-
-    expect(status.state).toBe('unauthorized')
-    expect(status.errorCode).toBe('emulator_adb_unauthorized')
-  })
-
-  it('disconnects the last-connected address when the RPC call omits one', async () => {
-    let connected = true
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary !== '/sdk/adb') {
-        return { stdout: '', stderr: '', code: 0 }
-      }
-      if (args[0] === 'connect') {
-        return { stdout: `connected to ${ADDRESS}`, stderr: '', code: 0 }
-      }
-      if (args[0] === 'disconnect') {
-        connected = false
-        return { stdout: `disconnected ${ADDRESS}`, stderr: '', code: 0 }
-      }
-      if (args.join(' ') === 'devices -l') {
-        return {
-          stdout: connected
-            ? `List of devices attached\n${ADDRESS}\tdevice`
-            : 'List of devices attached',
-          stderr: '',
-          code: 0
-        }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-    const cmds = commands(bridge)
-    await cmds.emulatorAdbConnect({ address: ADDRESS })
-
-    const status = await cmds.emulatorAdbDisconnect({})
-
-    expect(status).toEqual({ state: 'disconnected', address: ADDRESS, serial: null })
-  })
-
-  it('fails typed when disconnect has no address and nothing has ever connected', async () => {
-    const bridge = new EmulatorBridge()
-
-    await expect(commands(bridge).emulatorAdbDisconnect({})).rejects.toMatchObject({
-      code: 'emulator_adb_not_connected'
-    })
-    expect(androidCommandRunnerMock).not.toHaveBeenCalled()
-  })
-
-  it('reports disconnected for a status call with no address and nothing ever connected, without any adb I/O', async () => {
-    const bridge = new EmulatorBridge()
-
-    const status = await commands(bridge).emulatorAdbConnectionStatus({})
-
-    expect(status).toEqual({ state: 'disconnected', address: null, serial: null })
-    expect(androidCommandRunnerMock).not.toHaveBeenCalled()
-  })
-
-  it('resolves two overlapping connect calls to a single underlying adb connect invocation', async () => {
-    let resolveConnect: (result: {
-      stdout: string
-      stderr: string
-      code: number
-    }) => void = () => {}
-    const connectGate = new Promise<{ stdout: string; stderr: string; code: number }>((resolve) => {
-      resolveConnect = resolve
-    })
-    androidCommandRunnerMock.mockImplementation(async (binary: string, args: readonly string[]) => {
-      if (binary !== '/sdk/adb') {
-        return { stdout: '', stderr: '', code: 0 }
-      }
-      if (args[0] === 'connect') {
-        return connectGate
-      }
-      if (args.join(' ') === 'devices -l') {
-        return { stdout: `List of devices attached\n${ADDRESS}\tdevice`, stderr: '', code: 0 }
-      }
-      return { stdout: '', stderr: '', code: 0 }
-    })
-    const bridge = new EmulatorBridge()
-    const cmds = commands(bridge)
-
-    const first = cmds.emulatorAdbConnect({ address: ADDRESS })
-    const second = cmds.emulatorAdbConnect({ address: ADDRESS })
-    resolveConnect({ stdout: `connected to ${ADDRESS}`, stderr: '', code: 0 })
-    const [firstStatus, secondStatus] = await Promise.all([first, second])
-
-    expect(firstStatus).toEqual(secondStatus)
-    const connectCalls = androidCommandRunnerMock.mock.calls.filter(
-      (call) => (call[1] as string[])[0] === 'connect'
-    )
-    expect(connectCalls).toHaveLength(1)
   })
 })
